refactor(tv): move utils into correct folders and reword some fn's
