+++ projects/client/src/lib/features/mobile-os/handle.ts
@@ -1,18 +1,7 @@
+import { extractOS } from '$lib/utils/devices/extractOS.ts';
 import type { Handle } from '@sveltejs/kit';
 
 export const OS_PLACEHOLDER = '%mobile.os%';
-function extractMobileOS(agent: string): 'android' | 'ios' | 'unknown' {
-  // TODO: @seferturan, check if we can combine this with mobileAppleDeviceTriggerHack
-  if (/Android/i.test(agent)) {
-    return 'android';
-  }
-
-  if (/iPad|iPhone|iPod/.test(agent)) {
-    return 'ios';
-  }
-
-  return 'unknown';
-}
 
 export const handle: Handle = (
   { event, resolve },
@@ -23,7 +12,7 @@ export const handle: Handle = (
     transformPageChunk({ html, done }) {
       if (!done) return html;
       return html
-        .replace(OS_PLACEHOLDER, extractMobileOS(agent ?? ''));
+        .replace(OS_PLACEHOLDER, extractOS(agent ?? ''));
     },
   });
 };
