+++ projects/client/src/lib/utils/devices/isTV.spec.ts
@@ -1,25 +1,25 @@
 import { describe, expect, it } from 'vitest';
-import { isTvAgent } from './isTvAgent.ts';
+import { isTV } from './isTV.ts';
 
-describe('isTvAgent', () => {
+describe('isTV', () => {
   it('should return false when userAgent is null', () => {
-    expect(isTvAgent(null)).toBe(false);
+    expect(isTV(null)).toBe(false);
   });
 
   it('should return false when userAgent is undefined', () => {
-    expect(isTvAgent(undefined)).toBe(false);
+    expect(isTV(undefined)).toBe(false);
   });
 
   it('should return true for Android TV user agents', () => {
     const androidTvUserAgent =
       'Mozilla/5.0 (Linux; Android 9; ANDROIDTV DEVICE) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/80.0.3987.132 Safari/537.36';
-    expect(isTvAgent(androidTvUserAgent)).toBe(true);
+    expect(isTV(androidTvUserAgent)).toBe(true);
   });
 
   it('should return false for regular Android user agents without TV', () => {
     const androidUserAgent =
       'Mozilla/5.0 (Linux; Android 10; SM-A505F) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/89.0.4389.105 Mobile Safari/537.36';
-    expect(isTvAgent(androidUserAgent)).toBe(false);
+    expect(isTV(androidUserAgent)).toBe(false);
   });
 
   it('should return false for non-Android user agents', () => {
@@ -28,7 +28,7 @@ describe('isTvAgent', () => {
     const desktopUserAgent =
       'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36';
 
-    expect(isTvAgent(iosUserAgent)).toBe(false);
-    expect(isTvAgent(desktopUserAgent)).toBe(false);
+    expect(isTV(iosUserAgent)).toBe(false);
+    expect(isTV(desktopUserAgent)).toBe(false);
   });
 });
